criteria {rprojroot}	R Documentation

Prespecified criteria

Description

This is a collection of commonly used root criteria.

Usage

criteria

is_rstudio_project

is_r_package

is_remake_project

is_projectile_project

is_git_root

is_svn_root

is_vcs_root

is_testthat

from_wd

Format

An object of class root_criteria of length 9.

Details

is_rstudio_project looks for a file with extension .Rproj.

is_r_package looks for a DESCRIPTION file.

is_remake_project looks for a remake.yml file.

is_projectile_project looks for a .projectile file.

is_git_project looks for a .git directory.

is_svn_project looks for a .svn directory.

is_vcs_project looks for the root of a version control system, currently only Git and SVN are supported.

is_testthat looks for the testthat directory, works when developing, testing, and checking a package.

from_wd uses the current working directory.

[Package rprojroot version 1.3-2 Index]