Finding Files in Project Subdirectories

Robust, reliable and flexible paths to files below a project root. The 'root' of a project is defined as a directory that matches a certain criterion, e.g., it contains a certain regular file.

Details

See the "Value" section in root_criterion for documentation of root criterion objects, and criteria for useful predefined root criteria.

Examples

criteria

#> $is_rstudio_project
#> Root criterion: Contains a file matching '[.]Rproj$' with contents matching '^Version: ' in the first 1 lines
#> 
#> $is_r_package
#> Root criterion: Contains a file 'DESCRIPTION' with contents matching '^Package: '
#> 
#> $is_remake_project
#> Root criterion: Contains a file 'remake.yml'
#> 
#> $is_testthat
#> Root criterion: Directory name is 'testthat' (also look in subdirectories: 'tests/testthat', 'testthat')
#> 
#> $from_wd
#> Root criterion: From current working directory
#> 
#> attr(,"class")
#> [1] "root_criteria"

## Not run: ------------------------------------
# is_r_package$find_file("NAMESPACE")
# root_fun <- is_r_package$make_fix_file()
# root_fun("NAMESPACE")
## ---------------------------------------------