rprojroot-package {rprojroot}	R Documentation

rprojroot: Finding Files in Project Subdirectories

Description

Robust, reliable and flexible paths to files below a project root. The 'root' of a project is defined as a directory that matches a certain criterion, e.g., it contains a certain regular file.

Details

See the "Value" section in root_criterion() for documentation of root criterion objects, and criteria for useful predefined root criteria.

Author(s)

Maintainer: Kirill Müller kirill@cynkra.com (ORCID)

See Also

Useful links:

https://rprojroot.r-lib.org/

https://github.com/r-lib/rprojroot

Report bugs at https://github.com/r-lib/rprojroot/issues

Examples

criteria
## Not run: 
is_r_package$find_file("NAMESPACE")
root_fun <- is_r_package$make_fix_file()
root_fun("NAMESPACE")

## End(Not run)

[Package rprojroot version 2.0.4 Index]