rprojroot: Finding Files in Project Subdirectories

Robust, reliable and flexible paths to files below a project root. The 'root' of a project is defined as a directory that matches a certain criterion, e.g., it contains a certain regular file.

Details

See the "Value" section in root_criterion() for documentation of root criterion objects, and criteria() for useful predefined root criteria.

See also

Useful links:

https://github.com/krlmlr/rprojroot

https://krlmlr.github.io/rprojroot

Report bugs at https://github.com/krlmlr/rprojroot/issues

Examples

# NOT RUN {
criteria
# }# NOT RUN {
is_r_package$find_file("NAMESPACE")
root_fun <- is_r_package$make_fix_file()
root_fun("NAMESPACE")
# }